utils/checkports/requirements.txt
GitPython==3.1.31
python3-nmap==1.6.0
